% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.exportFormats}
\alias{About.exportFormats}
\title{About.exportFormats Object}
\usage{
About.exportFormats()
}
\value{
About.exportFormats object
}
\description{
About.exportFormats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of source MIME type to possible targets for all supported exports.
}
\seealso{
Other About functions: \code{\link{About.importFormats}},
  \code{\link{About.maxImportSizes}},
  \code{\link{About.storageQuota}}, \code{\link{About}}
}
