% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.importFormats}
\alias{About.importFormats}
\title{About.importFormats Object}
\usage{
About.importFormats()
}
\value{
About.importFormats object
}
\description{
About.importFormats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of source MIME type to possible targets for all supported imports.
}
\seealso{
Other About functions: \code{\link{About.exportFormats}},
  \code{\link{About.maxImportSizes}},
  \code{\link{About.storageQuota}}, \code{\link{About}}
}

