% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.maxImportSizes}
\alias{About.maxImportSizes}
\title{About.maxImportSizes Object}
\usage{
About.maxImportSizes()
}
\value{
About.maxImportSizes object
}
\description{
About.maxImportSizes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of maximum import sizes by MIME type, in bytes.
}
\seealso{
Other About functions: \code{\link{About.exportFormats}},
  \code{\link{About.importFormats}},
  \code{\link{About.storageQuota}}, \code{\link{About}}
}

