% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About}
\alias{About}
\title{Drive API Objects 
Manages files in Drive including uploading, downloading, searching, detecting changes, and updating sharing permissions.}
\usage{
About(About.exportFormats = NULL, About.importFormats = NULL,
  About.maxImportSizes = NULL, About.storageQuota = NULL,
  appInstalled = NULL, exportFormats = NULL, folderColorPalette = NULL,
  importFormats = NULL, maxImportSizes = NULL, maxUploadSize = NULL,
  storageQuota = NULL, user = NULL)
}
\arguments{
\item{About.exportFormats}{The \link{About.exportFormats} object or list of objects}

\item{About.importFormats}{The \link{About.importFormats} object or list of objects}

\item{About.maxImportSizes}{The \link{About.maxImportSizes} object or list of objects}

\item{About.storageQuota}{The \link{About.storageQuota} object or list of objects}

\item{appInstalled}{Whether the user has installed the requesting app}

\item{exportFormats}{A map of source MIME type to possible targets for all supported exports}

\item{folderColorPalette}{The currently supported folder colors as RGB hex strings}

\item{importFormats}{A map of source MIME type to possible targets for all supported imports}

\item{maxImportSizes}{A map of maximum import sizes by MIME type, in bytes}

\item{maxUploadSize}{The maximum upload size in bytes}

\item{storageQuota}{The user's storage quota limits and usage}

\item{user}{The authenticated user}
}
\value{
About object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:49:53
filename: /Users/mark/dev/R/autoGoogleAPI/googledrivev3.auto/R/drive_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
About Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the user, the user's Drive, and system capabilities.
}
\seealso{
Other About functions: \code{\link{About.exportFormats}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxImportSizes}},
  \code{\link{About.storageQuota}}
}
