% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{About.storageQuota}
\alias{About.storageQuota}
\title{About.storageQuota Object}
\usage{
About.storageQuota(limit = NULL, usage = NULL, usageInDrive = NULL,
  usageInDriveTrash = NULL)
}
\arguments{
\item{limit}{The usage limit, if applicable}

\item{usage}{The total usage across all services}

\item{usageInDrive}{The usage by all files in Google Drive}

\item{usageInDriveTrash}{The usage by trashed files in Google Drive}
}
\value{
About.storageQuota object
}
\description{
About.storageQuota Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user's storage quota limits and usage. All fields are measured in bytes.
}
\seealso{
Other About functions: \code{\link{About.exportFormats}},
  \code{\link{About.importFormats}},
  \code{\link{About.maxImportSizes}}, \code{\link{About}}
}

