% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Change}
\alias{Change}
\title{Change Object}
\usage{
Change(file = NULL, fileId = NULL, removed = NULL, time = NULL)
}
\arguments{
\item{file}{The updated state of the file}

\item{fileId}{The ID of the file which has changed}

\item{removed}{Whether the file has been removed from the view of the changes list, for example by deletion or lost access}

\item{time}{The time of this change (RFC 3339 date-time)}
}
\value{
Change object
}
\description{
Change Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A change to a file.
}

