% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ChangeList}
\alias{ChangeList}
\title{ChangeList Object}
\usage{
ChangeList(changes = NULL, newStartPageToken = NULL, nextPageToken = NULL)
}
\arguments{
\item{changes}{The page of changes}

\item{newStartPageToken}{The starting page token for future changes}

\item{nextPageToken}{The page token for the next page of changes}
}
\value{
ChangeList object
}
\description{
ChangeList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of changes for a user.
}

