% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{changes.list}
\alias{changes.list}
\title{Lists changes for a user.}
\usage{
changes.list(pageToken, includeRemoved = NULL, pageSize = NULL,
  restrictToMyDrive = NULL, spaces = NULL)
}
\arguments{
\item{pageToken}{The token for continuing a previous list request on the next page}

\item{includeRemoved}{Whether to include changes indicating that items have left the view of the changes list, for example by deletion or lost access}

\item{pageSize}{The maximum number of changes to return per page}

\item{restrictToMyDrive}{Whether to restrict the results to changes inside the My Drive hierarchy}

\item{spaces}{A comma-separated list of spaces to query within the user corpus}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

