% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{changes.watch}
\alias{changes.watch}
\title{Subscribes to changes for a user.}
\usage{
changes.watch(Channel, pageToken, includeCorpusRemovals = NULL,
  includeRemoved = NULL, includeTeamDriveItems = NULL, pageSize = NULL,
  restrictToMyDrive = NULL, spaces = NULL, supportsTeamDrives = NULL,
  teamDriveId = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{pageToken}{The token for continuing a previous list request on the next page}

\item{includeCorpusRemovals}{Whether changes should include the file resource if the file is still accessible by the user at the time of the request, even when a file was removed from the list of changes and there will be no further change entries for this file}

\item{includeRemoved}{Whether to include changes indicating that items have been removed from the list of changes, for example by deletion or loss of access}

\item{includeTeamDriveItems}{Whether Team Drive files or changes should be included in results}

\item{pageSize}{The maximum number of changes to return per page}

\item{restrictToMyDrive}{Whether to restrict the results to changes inside the My Drive hierarchy}

\item{spaces}{A comma-separated list of spaces to query within the user corpus}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}

\item{teamDriveId}{The Team Drive from which changes will be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.metadata.readonly, https://www.googleapis.com/auth/drive.photos.readonly, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}}, \code{\link{channels.stop}},
  \code{\link{files.watch}}
}
