% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Channel}
\alias{Channel}
\title{Channel Object}
\usage{
Channel(Channel.params = NULL, address = NULL, expiration = NULL,
  id = NULL, params = NULL, payload = NULL, resourceId = NULL,
  resourceUri = NULL, token = NULL, type = NULL)
}
\arguments{
\item{Channel.params}{The \link{Channel.params} object or list of objects}

\item{address}{The address where notifications are delivered for this channel}

\item{expiration}{Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds}

\item{id}{A UUID or similar unique string that identifies this channel}

\item{params}{Additional parameters controlling delivery channel behavior}

\item{payload}{A Boolean value to indicate whether payload is wanted}

\item{resourceId}{An opaque ID that identifies the resource being watched on this channel}

\item{resourceUri}{A version-specific identifier for the watched resource}

\item{token}{An arbitrary string delivered to the target address with each notification delivered over this channel}

\item{type}{The type of delivery mechanism used for this channel}
}
\value{
Channel object
}
\description{
Channel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An notification channel used to watch for resource changes.
}
\seealso{
Other Channel functions: \code{\link{Channel.params}},
  \code{\link{changes.watch}}, \code{\link{channels.stop}},
  \code{\link{files.watch}}
}

