% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Comment.quotedFileContent}
\alias{Comment.quotedFileContent}
\title{Comment.quotedFileContent Object}
\usage{
Comment.quotedFileContent(mimeType = NULL, value = NULL)
}
\arguments{
\item{mimeType}{The MIME type of the quoted content}

\item{value}{The quoted content itself}
}
\value{
Comment.quotedFileContent object
}
\description{
Comment.quotedFileContent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The file content to which the comment refers, typically within the anchor region. For a text file, for example, this would be the text at the location of the comment.
}
\seealso{
Other Comment functions: \code{\link{Comment}},
  \code{\link{comments.create}},
  \code{\link{comments.update}}
}
