% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Comment}
\alias{Comment}
\title{Comment Object}
\usage{
Comment(Comment.quotedFileContent = NULL, anchor = NULL, author = NULL,
  content = NULL, createdTime = NULL, deleted = NULL,
  htmlContent = NULL, id = NULL, modifiedTime = NULL,
  quotedFileContent = NULL, replies = NULL, resolved = NULL)
}
\arguments{
\item{Comment.quotedFileContent}{The \link{Comment.quotedFileContent} object or list of objects}

\item{anchor}{A region of the document represented as a JSON string}

\item{author}{The user who created the comment}

\item{content}{The plain text content of the comment}

\item{createdTime}{The time at which the comment was created (RFC 3339 date-time)}

\item{deleted}{Whether the comment has been deleted}

\item{htmlContent}{The content of the comment with HTML formatting}

\item{id}{The ID of the comment}

\item{modifiedTime}{The last time the comment or any of its replies was modified (RFC 3339 date-time)}

\item{quotedFileContent}{The file content to which the comment refers, typically within the anchor region}

\item{replies}{The full list of replies to the comment in chronological order}

\item{resolved}{Whether the comment has been resolved by one of its replies}
}
\value{
Comment object
}
\description{
Comment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A comment on a file.
}
\seealso{
Other Comment functions: \code{\link{Comment.quotedFileContent}},
  \code{\link{comments.create}},
  \code{\link{comments.update}}
}
