% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{CommentList}
\alias{CommentList}
\title{CommentList Object}
\usage{
CommentList(comments = NULL, nextPageToken = NULL)
}
\arguments{
\item{comments}{The page of comments}

\item{nextPageToken}{The page token for the next page of comments}
}
\value{
CommentList object
}
\description{
CommentList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of comments on a file.
}

