% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{comments.create}
\alias{comments.create}
\title{Creates a new comment on a file.}
\usage{
comments.create(Comment, fileId)
}
\arguments{
\item{Comment}{The \link{Comment} object to pass to this method}

\item{fileId}{The ID of the file}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Comment functions: \code{\link{Comment.quotedFileContent}},
  \code{\link{Comment}}, \code{\link{comments.update}}
}

