% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{comments.list}
\alias{comments.list}
\title{Lists a file's comments.}
\usage{
comments.list(fileId, includeDeleted = NULL, pageSize = NULL,
  pageToken = NULL, startModifiedTime = NULL)
}
\arguments{
\item{fileId}{The ID of the file}

\item{includeDeleted}{Whether to include deleted comments}

\item{pageSize}{The maximum number of comments to return per page}

\item{pageToken}{The token for continuing a previous list request on the next page}

\item{startModifiedTime}{The minimum value of 'modifiedTime' for the result comments (RFC 3339 date-time)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}

