% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\docType{package}
\name{drive_googleAuthR}
\alias{drive_googleAuthR}
\alias{drive_googleAuthR-package}
\title{Drive API
Manages files in Drive including uploading, downloading, searching, detecting changes, and updating sharing permissions.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:49:52
filename: /Users/mark/dev/R/autoGoogleAPI/googledrivev3.auto/R/drive_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/drive.scripts
}
}
