% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.appProperties}
\alias{File.appProperties}
\title{File.appProperties Object}
\usage{
File.appProperties()
}
\value{
File.appProperties object
}
\description{
File.appProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of arbitrary key-value pairs which are private to the requesting app.Entries with null values are cleared in update and copy requests.
}
\seealso{
Other File functions: \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}

