% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.contentHints}
\alias{File.contentHints}
\title{File.contentHints Object}
\usage{
File.contentHints(File.contentHints.thumbnail = NULL, indexableText = NULL,
  thumbnail = NULL)
}
\arguments{
\item{File.contentHints.thumbnail}{The \link{File.contentHints.thumbnail} object or list of objects}

\item{indexableText}{Text to be indexed for the file to improve fullText queries}

\item{thumbnail}{A thumbnail for the file}
}
\value{
File.contentHints object
}
\description{
File.contentHints Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information about the content of the file. These fields are never populated in responses.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}

