% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.contentHints.thumbnail}
\alias{File.contentHints.thumbnail}
\title{File.contentHints.thumbnail Object}
\usage{
File.contentHints.thumbnail(image = NULL, mimeType = NULL)
}
\arguments{
\item{image}{The thumbnail data encoded with URL-safe Base64 (RFC 4648 section 5)}

\item{mimeType}{The MIME type of the thumbnail}
}
\value{
File.contentHints.thumbnail object
}
\description{
File.contentHints.thumbnail Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A thumbnail for the file. This will only be used if Drive cannot generate a standard thumbnail.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}

