% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.imageMediaMetadata}
\alias{File.imageMediaMetadata}
\title{File.imageMediaMetadata Object}
\usage{
File.imageMediaMetadata(File.imageMediaMetadata.location = NULL,
  aperture = NULL, cameraMake = NULL, cameraModel = NULL,
  colorSpace = NULL, exposureBias = NULL, exposureMode = NULL,
  exposureTime = NULL, flashUsed = NULL, focalLength = NULL,
  height = NULL, isoSpeed = NULL, lens = NULL, location = NULL,
  maxApertureValue = NULL, meteringMode = NULL, rotation = NULL,
  sensor = NULL, subjectDistance = NULL, time = NULL,
  whiteBalance = NULL, width = NULL)
}
\arguments{
\item{File.imageMediaMetadata.location}{The \link{File.imageMediaMetadata.location} object or list of objects}

\item{aperture}{The aperture used to create the photo (f-number)}

\item{cameraMake}{The make of the camera used to create the photo}

\item{cameraModel}{The model of the camera used to create the photo}

\item{colorSpace}{The color space of the photo}

\item{exposureBias}{The exposure bias of the photo (APEX value)}

\item{exposureMode}{The exposure mode used to create the photo}

\item{exposureTime}{The length of the exposure, in seconds}

\item{flashUsed}{Whether a flash was used to create the photo}

\item{focalLength}{The focal length used to create the photo, in millimeters}

\item{height}{The height of the image in pixels}

\item{isoSpeed}{The ISO speed used to create the photo}

\item{lens}{The lens used to create the photo}

\item{location}{Geographic location information stored in the image}

\item{maxApertureValue}{The smallest f-number of the lens at the focal length used to create the photo (APEX value)}

\item{meteringMode}{The metering mode used to create the photo}

\item{rotation}{The rotation in clockwise degrees from the image's original orientation}

\item{sensor}{The type of sensor used to create the photo}

\item{subjectDistance}{The distance to the subject of the photo, in meters}

\item{time}{The date and time the photo was taken (EXIF DateTime)}

\item{whiteBalance}{The white balance mode used to create the photo}

\item{width}{The width of the image in pixels}
}
\value{
File.imageMediaMetadata object
}
\description{
File.imageMediaMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional metadata about image media, if available.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}
