% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.properties}
\alias{File.properties}
\title{File.properties Object}
\usage{
File.properties()
}
\value{
File.properties object
}
\description{
File.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of arbitrary key-value pairs which are visible to all apps.Entries with null values are cleared in update and copy requests.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}

