% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File}
\alias{File}
\title{File Object}
\usage{
File(File.appProperties = NULL, File.capabilities = NULL,
  File.contentHints = NULL, File.contentHints.thumbnail = NULL,
  File.imageMediaMetadata = NULL, File.imageMediaMetadata.location = NULL,
  File.properties = NULL, File.videoMediaMetadata = NULL,
  appProperties = NULL, capabilities = NULL, contentHints = NULL,
  createdTime = NULL, description = NULL, explicitlyTrashed = NULL,
  fileExtension = NULL, folderColorRgb = NULL, fullFileExtension = NULL,
  headRevisionId = NULL, iconLink = NULL, id = NULL,
  imageMediaMetadata = NULL, isAppAuthorized = NULL,
  lastModifyingUser = NULL, md5Checksum = NULL, mimeType = NULL,
  modifiedByMeTime = NULL, modifiedTime = NULL, name = NULL,
  originalFilename = NULL, ownedByMe = NULL, owners = NULL,
  parents = NULL, permissions = NULL, properties = NULL,
  quotaBytesUsed = NULL, shared = NULL, sharedWithMeTime = NULL,
  sharingUser = NULL, size = NULL, spaces = NULL, starred = NULL,
  thumbnailLink = NULL, trashed = NULL, version = NULL,
  videoMediaMetadata = NULL, viewedByMe = NULL, viewedByMeTime = NULL,
  viewersCanCopyContent = NULL, webContentLink = NULL, webViewLink = NULL,
  writersCanShare = NULL)
}
\arguments{
\item{File.appProperties}{The \link{File.appProperties} object or list of objects}

\item{File.capabilities}{The \link{File.capabilities} object or list of objects}

\item{File.contentHints}{The \link{File.contentHints} object or list of objects}

\item{File.contentHints.thumbnail}{The \link{File.contentHints.thumbnail} object or list of objects}

\item{File.imageMediaMetadata}{The \link{File.imageMediaMetadata} object or list of objects}

\item{File.imageMediaMetadata.location}{The \link{File.imageMediaMetadata.location} object or list of objects}

\item{File.properties}{The \link{File.properties} object or list of objects}

\item{File.videoMediaMetadata}{The \link{File.videoMediaMetadata} object or list of objects}

\item{appProperties}{A collection of arbitrary key-value pairs which are private to the requesting app}

\item{capabilities}{Capabilities the current user has on the file}

\item{contentHints}{Additional information about the content of the file}

\item{createdTime}{The time at which the file was created (RFC 3339 date-time)}

\item{description}{A short description of the file}

\item{explicitlyTrashed}{Whether the file has been explicitly trashed, as opposed to recursively trashed from a parent folder}

\item{fileExtension}{The final component of fullFileExtension}

\item{folderColorRgb}{The color for a folder as an RGB hex string}

\item{fullFileExtension}{The full file extension extracted from the name field}

\item{headRevisionId}{The ID of the file's head revision}

\item{iconLink}{A static, unauthenticated link to the file's icon}

\item{id}{The ID of the file}

\item{imageMediaMetadata}{Additional metadata about image media, if available}

\item{isAppAuthorized}{Whether the file was created or opened by the requesting app}

\item{lastModifyingUser}{The last user to modify the file}

\item{md5Checksum}{The MD5 checksum for the content of the file}

\item{mimeType}{The MIME type of the file}

\item{modifiedByMeTime}{The last time the file was modified by the user (RFC 3339 date-time)}

\item{modifiedTime}{The last time the file was modified by anyone (RFC 3339 date-time)}

\item{name}{The name of the file}

\item{originalFilename}{The original filename of the uploaded content if available, or else the original value of the name field}

\item{ownedByMe}{Whether the user owns the file}

\item{owners}{The owners of the file}

\item{parents}{The IDs of the parent folders which contain the file}

\item{permissions}{The full list of permissions for the file}

\item{properties}{A collection of arbitrary key-value pairs which are visible to all apps}

\item{quotaBytesUsed}{The number of storage quota bytes used by the file}

\item{shared}{Whether the file has been shared}

\item{sharedWithMeTime}{The time at which the file was shared with the user, if applicable (RFC 3339 date-time)}

\item{sharingUser}{The user who shared the file with the requesting user, if applicable}

\item{size}{The size of the file's content in bytes}

\item{spaces}{The list of spaces which contain the file}

\item{starred}{Whether the user has starred the file}

\item{thumbnailLink}{A short-lived link to the file's thumbnail, if available}

\item{trashed}{Whether the file has been trashed, either explicitly or from a trashed parent folder}

\item{version}{A monotonically increasing version number for the file}

\item{videoMediaMetadata}{Additional metadata about video media}

\item{viewedByMe}{Whether the file has been viewed by this user}

\item{viewedByMeTime}{The last time the file was viewed by the user (RFC 3339 date-time)}

\item{viewersCanCopyContent}{Whether users with only reader or commenter permission can copy the file's content}

\item{webContentLink}{A link for downloading the content of the file in a browser}

\item{webViewLink}{A link for opening the file in a relevant Google editor or viewer in a browser}

\item{writersCanShare}{Whether users with only writer permission can modify the file's permissions}
}
\value{
File object
}
\description{
File Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata for a file.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{files.copy}}, \code{\link{files.create}},
  \code{\link{files.update}}
}

