% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{FileList}
\alias{FileList}
\title{FileList Object}
\usage{
FileList(files = NULL, incompleteSearch = NULL, nextPageToken = NULL)
}
\arguments{
\item{files}{The list of files}

\item{incompleteSearch}{Whether the search process was incomplete}

\item{nextPageToken}{The page token for the next page of files}
}
\value{
FileList object
}
\description{
FileList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of files.
}
