% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.copy}
\alias{files.copy}
\title{Creates a copy of a file and applies any requested updates with patch semantics.}
\usage{
files.copy(File, fileId, ignoreDefaultVisibility = NULL,
  keepRevisionForever = NULL, ocrLanguage = NULL,
  supportsTeamDrives = NULL)
}
\arguments{
\item{File}{The \link{File} object to pass to this method}

\item{fileId}{The ID of the file}

\item{ignoreDefaultVisibility}{Whether to ignore the domain's default visibility settings for the created file}

\item{keepRevisionForever}{Whether to set the 'keepForever' field in the new head revision}

\item{ocrLanguage}{A language hint for OCR processing during image import (ISO 639-1 code)}

\item{supportsTeamDrives}{Whether the requesting application supports Team Drives}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.photos.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.photos.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.create}},
  \code{\link{files.update}}
}
