% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.export}
\alias{files.export}
\title{Exports a Google Doc to the requested MIME type and returns the exported content.}
\usage{
files.export(fileId, mimeType)
}
\arguments{
\item{fileId}{The ID of the file}

\item{mimeType}{The MIME type of the format requested for this export}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
