% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.generateIds}
\alias{files.generateIds}
\title{Generates a set of file IDs which can be provided in create requests.}
\usage{
files.generateIds(count = NULL, space = NULL)
}
\arguments{
\item{count}{The number of IDs to return}

\item{space}{The space in which the IDs can be used to create new files}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
