% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{files.update}
\alias{files.update}
\title{Updates a file's metadata and/or content with patch semantics.}
\usage{
files.update(File, fileId, addParents = NULL, keepRevisionForever = NULL,
  ocrLanguage = NULL, removeParents = NULL,
  useContentAsIndexableText = NULL)
}
\arguments{
\item{File}{The \link{File} object to pass to this method}

\item{fileId}{The ID of the file}

\item{addParents}{A comma-separated list of parent IDs to add}

\item{keepRevisionForever}{Whether to set the 'keepForever' field in the new head revision}

\item{ocrLanguage}{A language hint for OCR processing during image import (ISO 639-1 code)}

\item{removeParents}{A comma-separated list of parent IDs to remove}

\item{useContentAsIndexableText}{Whether to use the uploaded content as indexable text}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.scripts
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.appdata, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.metadata, https://www.googleapis.com/auth/drive.scripts)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.capabilities}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}
}

