% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{GeneratedIds}
\alias{GeneratedIds}
\title{GeneratedIds Object}
\usage{
GeneratedIds(ids = NULL, space = NULL)
}
\arguments{
\item{ids}{The IDs generated for the requesting user in the specified space}

\item{space}{The type of file that can be created with these IDs}
}
\value{
GeneratedIds object
}
\description{
GeneratedIds Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of generated file IDs which can be provided in create requests.
}
