% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Permission}
\alias{Permission}
\title{Permission Object}
\usage{
Permission(allowFileDiscovery = NULL, displayName = NULL, domain = NULL,
  emailAddress = NULL, expirationTime = NULL, id = NULL,
  photoLink = NULL, role = NULL, type = NULL)
}
\arguments{
\item{allowFileDiscovery}{Whether the permission allows the file to be discovered through search}

\item{displayName}{A displayable name for users, groups or domains}

\item{domain}{The domain to which this permission refers}

\item{emailAddress}{The email address of the user or group to which this permission refers}

\item{expirationTime}{The time at which this permission will expire (RFC 3339 date-time)}

\item{id}{The ID of this permission}

\item{photoLink}{A link to the user's profile photo, if available}

\item{role}{The role granted by this permission}

\item{type}{The type of the grantee}
}
\value{
Permission object
}
\description{
Permission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A permission for a file. A permission grants a user, group, domain or the world access to a file or a folder hierarchy.
}
\seealso{
Other Permission functions: \code{\link{permissions.create}},
  \code{\link{permissions.update}}
}

