% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Permission.teamDrivePermissionDetails}
\alias{Permission.teamDrivePermissionDetails}
\title{Permission.teamDrivePermissionDetails Object}
\usage{
Permission.teamDrivePermissionDetails()
}
\value{
Permission.teamDrivePermissionDetails object
}
\description{
Permission.teamDrivePermissionDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Permission functions: \code{\link{Permission}},
  \code{\link{permissions.create}},
  \code{\link{permissions.update}}
}
