% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{PermissionList}
\alias{PermissionList}
\title{PermissionList Object}
\usage{
PermissionList(nextPageToken = NULL, permissions = NULL)
}
\arguments{
\item{nextPageToken}{The page token for the next page of permissions}

\item{permissions}{The list of permissions}
}
\value{
PermissionList object
}
\description{
PermissionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of permissions for a file.
}
