% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{permissions.create}
\alias{permissions.create}
\title{Creates a permission for a file.}
\usage{
permissions.create(Permission, fileId, emailMessage = NULL,
  sendNotificationEmail = NULL, transferOwnership = NULL)
}
\arguments{
\item{Permission}{The \link{Permission} object to pass to this method}

\item{fileId}{The ID of the file}

\item{emailMessage}{A custom message to include in the notification email}

\item{sendNotificationEmail}{Whether to send a notification email when sharing to users or groups}

\item{transferOwnership}{Whether to transfer ownership to the specified user and downgrade the current owner to a writer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other Permission functions: \code{\link{Permission}},
  \code{\link{permissions.update}}
}

