% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{ReplyList}
\alias{ReplyList}
\title{ReplyList Object}
\usage{
ReplyList(nextPageToken = NULL, replies = NULL)
}
\arguments{
\item{nextPageToken}{The page token for the next page of replies}

\item{replies}{The page of replies}
}
\value{
ReplyList object
}
\description{
ReplyList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of replies to a comment on a file.
}

