% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{Revision}
\alias{Revision}
\title{Revision Object}
\usage{
Revision(id = NULL, keepForever = NULL, lastModifyingUser = NULL,
  md5Checksum = NULL, mimeType = NULL, modifiedTime = NULL,
  originalFilename = NULL, publishAuto = NULL, published = NULL,
  publishedOutsideDomain = NULL, size = NULL)
}
\arguments{
\item{id}{The ID of the revision}

\item{keepForever}{Whether to keep this revision forever, even if it is no longer the head revision}

\item{lastModifyingUser}{The last user to modify this revision}

\item{md5Checksum}{The MD5 checksum of the revision's content}

\item{mimeType}{The MIME type of the revision}

\item{modifiedTime}{The last time the revision was modified (RFC 3339 date-time)}

\item{originalFilename}{The original filename used to create this revision}

\item{publishAuto}{Whether subsequent revisions will be automatically republished}

\item{published}{Whether this revision is published}

\item{publishedOutsideDomain}{Whether this revision is published outside the domain}

\item{size}{The size of the revision's content in bytes}
}
\value{
Revision object
}
\description{
Revision Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata for a revision to a file.
}
\seealso{
Other Revision functions: \code{\link{revisions.update}}
}

