% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_functions.R
\name{team.update}
\alias{team.update}
\title{Updates a Team Drive's metadata}
\usage{
team.update(TeamDrive, teamDriveId)
}
\arguments{
\item{TeamDrive}{The \link{TeamDrive} object to pass to this method}

\item{teamDriveId}{The ID of the Team Drive}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}

Other TeamDrive functions: \code{\link{TeamDrive.capabilities}},
  \code{\link{TeamDrive}}, \code{\link{team.create}}
}
