% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{TeamDrive.capabilities}
\alias{TeamDrive.capabilities}
\title{TeamDrive.capabilities Object}
\usage{
TeamDrive.capabilities(canAddChildren = NULL, canComment = NULL,
  canCopy = NULL, canDeleteTeamDrive = NULL, canDownload = NULL,
  canEdit = NULL, canListChildren = NULL, canManageMembers = NULL,
  canReadRevisions = NULL, canRemoveChildren = NULL, canRename = NULL,
  canRenameTeamDrive = NULL, canShare = NULL)
}
\arguments{
\item{canAddChildren}{Whether the current user can add children to folders in this Team Drive}

\item{canComment}{Whether the current user can comment on files in this Team Drive}

\item{canCopy}{Whether files in this Team Drive can be copied by the current user}

\item{canDeleteTeamDrive}{Whether this Team Drive can be deleted by the current user}

\item{canDownload}{Whether files in this Team Drive can be downloaded by the current user}

\item{canEdit}{Whether files in this Team Drive can be edited by the current user}

\item{canListChildren}{Whether the current user can list the children of folders in this Team Drive}

\item{canManageMembers}{Whether the current user can add members to this Team Drive or remove them or change their role}

\item{canReadRevisions}{Whether the current user has read access to the Revisions resource of files in this Team Drive}

\item{canRemoveChildren}{Whether the current user can remove children from folders in this Team Drive}

\item{canRename}{Whether files or folders in this Team Drive can be renamed by the current user}

\item{canRenameTeamDrive}{Whether this Team Drive can be renamed by the current user}

\item{canShare}{Whether the current user can share files or folders in this Team Drive}
}
\value{
TeamDrive.capabilities object
}
\description{
TeamDrive.capabilities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Capabilities the current user has on this Team Drive.
}
\seealso{
Other TeamDrive functions: \code{\link{TeamDrive}},
  \code{\link{team.create}}, \code{\link{team.update}}
}
