% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{TeamDrive}
\alias{TeamDrive}
\title{TeamDrive Object}
\usage{
TeamDrive(TeamDrive.capabilities = NULL, capabilities = NULL, id = NULL,
  name = NULL)
}
\arguments{
\item{TeamDrive.capabilities}{The \link{TeamDrive.capabilities} object or list of objects}

\item{capabilities}{Capabilities the current user has on this Team Drive}

\item{id}{The ID of this Team Drive which is also the ID of the top level folder for this Team Drive}

\item{name}{The name of this Team Drive}
}
\value{
TeamDrive object
}
\description{
TeamDrive Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of a Team Drive.
}
\seealso{
Other TeamDrive functions: \code{\link{TeamDrive.capabilities}},
  \code{\link{team.create}}, \code{\link{team.update}}
}
