% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{TeamDriveList}
\alias{TeamDriveList}
\title{TeamDriveList Object}
\usage{
TeamDriveList(nextPageToken = NULL, teamDrives = NULL)
}
\arguments{
\item{nextPageToken}{The page token for the next page of Team Drives}

\item{teamDrives}{The list of Team Drives}
}
\value{
TeamDriveList object
}
\description{
TeamDriveList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Team Drives.
}
