% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(displayName = NULL, emailAddress = NULL, me = NULL,
  permissionId = NULL, photoLink = NULL)
}
\arguments{
\item{displayName}{A plain text displayable name for this user}

\item{emailAddress}{The email address of the user}

\item{me}{Whether this user is the requesting user}

\item{permissionId}{The user's ID as visible in Permission resources}

\item{photoLink}{A link to the user's profile photo, if available}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a Drive user.
}
