% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.capabilities}
\alias{File.capabilities}
\title{File.capabilities Object}
\usage{
File.capabilities(canComment = NULL, canCopy = NULL, canEdit = NULL,
  canReadRevisions = NULL, canShare = NULL)
}
\arguments{
\item{canComment}{Whether the user can comment on the file}

\item{canCopy}{Whether the user can copy the file}

\item{canEdit}{Whether the user can edit the file's content}

\item{canReadRevisions}{Whether the current user has read access to the Revisions resource of the file}

\item{canShare}{Whether the user can modify the file's permissions and sharing settings}
}
\value{
File.capabilities object
}
\description{
File.capabilities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Capabilities the current user has on the file.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.contentHints.thumbnail}},
  \code{\link{File.contentHints}},
  \code{\link{File.imageMediaMetadata.location}},
  \code{\link{File.imageMediaMetadata}},
  \code{\link{File.properties}},
  \code{\link{File.videoMediaMetadata}},
  \code{\link{File}}, \code{\link{files.copy}},
  \code{\link{files.create}}, \code{\link{files.update}}
}

