% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{RevisionList}
\alias{RevisionList}
\title{RevisionList Object}
\usage{
RevisionList(revisions = NULL)
}
\arguments{
\item{revisions}{The full list of revisions}
}
\value{
RevisionList object
}
\description{
RevisionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of revisions of a file.
}

